put ugly date parsing machinery into InstantUtils for better testability
All the hairy logic for parsing instants and creating Date (or
Calendar or Timestamp) objects from the result is now in static
methods in InstantUtils. This makes them easier to reuse and to test.

Testing revealed just how hacked-up and broken my earlier attempt to
provide #inst handler was. (Would it be a good time to mention that
Java's native date/time APIs are a disgrace?)

InstantUtils parses RFC3339 strings to ParsedInstant values. If you
wish to provide a custom Handler for #inst, extend
AbstractInstantHandler and implement translate(ParsedInstant).