Scanner builds on Reader directly
Instead of introducing our own abstract (Input), we've decided to just
build on Reader. It's more general than what we need, but it's less
code to maintain.

Also, Scanner really doesn't need to be visible to API users. I've pulled
it into the parser package, and made it package visible.