provide support for simplistic "protocols"
Here, I mean protocols in the Clojure/Objective-C sense: the ability
to "add" methods to a class after the fact. This is just the sort of
thing that we'll need in order to implement printing of edn values.

I fought with Java Generics for few hours, trying to figure out how
to convince them to DWIM, before finally giving up and just leaving
all the types raw. Hopefully no great loss.