remove IOException from the public interfaces of Printer, Parser
Printer and Parser no longer throw checked exceptions. Instead, they
may throw EdnIOException, which is a subclass of EdnException (a
RuntimeExcetption) that always wraps a cause that is an IOException.

Internally, IOExceptions are still used, meaning that the translation
to runtime exceptions takes place at the public interface, not at the
point where we read form the underlying reader or write to the
underlying writer.

Also, Scanner and Parser both try to make sure that the underlying
resource gets closed in case of an exception. This is redundant,
but does no harm apart from the unnecessary code.