Teach scanner to work with anything that is Readable
Previously it required an entire reader.

This complicates the internals of getting the next character
considerably since we end up doing our own buffer management
with a pair of CharBuffers which are swapped every time the
current read position moves form the first to the second buffer.
The only virtue of this implementation is that unit tests do still
pass.

(Patches for nicer solutions welcome.)