first step toward building Scanner on a PushbackReader
The hope is that, like clojure, I'll be able to confine the mutable
state of the Parse to the PushbackReader.

That's going to mean threading curr as a parameter through all
methods. To go really stateless, I'll have to thread pbr through as
well. I'll also have to deal with the fact that you can't unread
the end-of-stream marker.  Maybe this won't be such a problem; the
documentation of Reader.read() leaves open the possibility of calling
read() as often as you like past the end of the stream (and just
getting -1 back every time.)