save time with check on position and limit rather than try/catch
Initially, I thought that the common case of buff.get() working as
intended would be so much more common than the rare case of this
resulting in a BufferUnderflowException that it would make sense
to handle the BufferUnderflowException rather than checking that
position() < limit() before making the call to get().

As it turns out, "look before you leap" is faster here, even
for large buffers. (CombinedBenchmark reduced from 41.1 ms to 40.3 ms)