make Protocol generic; no longer limited to just Function
The domain of a Protocol can now be any time F, not just Function. The
interface Function has been removed.

You can think of the Protocol as a kind of immutable Map keyed by
Class, where lookups are resolved using the C3 learization when an
exact match for a given class can not be found in the map.