simplify Printer using newly generic Protocol
The only "configuration" required by Printer is the Protocol of
Printer.Fn implementations. Accordingly, we've removed Printer.Config
and Printer.Config.Buidler in favor of directly exposing Protocol and
Protocol.Builder.

This may need to change should Printer grow options, but as this
Printer isn't a pretty-printer, I don't anticipate that. A
pretty-printer would be a desirable option, but a design for it has
not cristalized yet.