#29: make master content identical with 0.4.0 tag
Edn-java is switching to git-flow conventions for branch management.
Normally, this would mean resetting master to the most recent release
tag, but it doesn't seem possible to remove and replace master on
github since it always counts as the branch that is currently checked
out.

The switch to git-flow entails:

- future commits on master will only be actual releases (not SNAPSHOT)
- the develop branch takes over the role currently played by 'master',
  but will build as version develop-SNAPSHOT.
- releases will be prepared on temporary release branches originating
  from develop, with merge back to develop and master.
- we decide on the release branch which version the release in
  preparation will have.
- merging a release branch to master is making a release. The resulting
  revision will be tagged accordingly.
- feature branches originate on develop and merge back to develop.