DefaultMapFactory should throw EdnSyntaxException
when there is not a value for every key instead of throwing
IllegalStateException since callers of the Parser should be expecting
syntax errors to be reported as EdnSyntaxException and this certainly
qualifies as a syntax error.

Also, we'll be adding another case where EdnSyntaxException can
be thrown: upon duplicate in maps, and duplicate elements in sets.