Define a stable automatic module name.
@brunchboy would like to modularize the project
https://github.com/Deep-Symmetry/lib-carabiner
but is unable to do so because edn-java does not define an automatic
module name. See
http://branchandbound.net/blog/java/2017/12/automatic-module-name/
for details on why this is necessary. All the other lib-carabiner
dependencies already do this, so as soon as edn-java does so, it will
be possible.

This pull request sets `us.bpsm.edn` as the automatic module name,
which makes sense as it is likely the module name you will use if you
ever decide to modularize this project itself.

Thanks for considering this change.