60: support \uXXXX style 16-bit unicode character literals
Atomic literal characters values in EDN may be written as using a
\u followed by exactly four hex digits to ultimately represent
a single 16-bit Java Character.

The hex digits are case insensitive, so 0123456789abcdefABCDEF are
all valid digits.

Characters outside of the the Basic Multilingual Pane can not be
expressed in this way and can not be represented in the Java virtual
machine as Characters. (Integers would have to be used, which would
require more substantial changes to edn-java.)