62: Accept symbols which contain #
Previously when we encountered the a '#' while parsing the name of a
symbol, we considered the symbol name ended and treated '#' as a
dispatch character.

We also correct printing of symbols immediately followed by a
set (which begins with #). These need to separated by whitespace in
order to parse correctly.

As a result

  [foo#bar {}]

Would parse as a vector of two elements:

 1. the symbol foo
 2. the empty map tagged with the symbol bar.

The correct parse would is:

 1. the symbol foo#bar
 2. the empty map.

Correcting this uncovered a round-tripping bug related to how sets are
printed:

 [foo #{}]

Parses initially as the symbol named 'foo' followed by empty
set. However, we print it as:

 [foo#{}]

And this, in turn parses as the symbol named 'foo#' followed by the
empty map.

Printing needs to preserve the space between foo and #{ in order for
this to be parsed correctly.